requirements.txt
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
 #optional_for_gpu_training
matplotlib==3.9.1
seaborn==0.13.2
torch==2.3.1
torchvision==0.18.1
scikit-learn==1.5.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0